Latest version of SpoiledApps (as of 3/16/2020)
You can download the folder right away and open it directly in Android Studio with no issue. This iteration contains a login screen, a registration screen and a home page with the bottom navigation bar. I did not modify the bottom navigation bar home page yet as I'd let you guys first get a feel and come to an agreement with the progress made so far!