Version 0.2 (as of 3/17/2020) Update
Hi all! This update is minor. Now includes the ability to click the register button on login page which then takes you to registration page! You can try it out yourselves on any Android device and should work fine as I did it myself and I had no issues! :) Also, I am going to have to start uploading my files as a .zip because GitHub has some limit of 100 files or less and the project is definitely LARGER than 100. So I just zipped it to beat their system. No biggie.