Database and Authenication ALL READY to go! :)
Hey everyone, I managed to fix the database issue, very minor issue! The registration page works PERFECTLY! Now, I'm pretty sure we discussed adding a vertifcation page after a new user signs up but we can work on that of course. Login page works somewhat because it currently does not check the credetials being used to sign into the app if its legit data of an account or not. Basically, it will log EVERYONE in. Overall though, the registration page is done, a verfication page is simple to add, and adding a check at login is simple enough too. 2 people could work on that. The bulk of our work now should be on the actual home page of the app which needs a LOT of work that we all are capable of doing of course! Anyways, we're literally 45% done with this project! Woohoo!!