Fixed a bug! (as of 5/5/2020)
Fixed a bug where once you register and are taken to the verification page,if you click on the "Take me back to login page" button, it will actually take you to the HomePage. This is because the app thinks you're signed in once you register for an account. This has been fixed so no worries :) If someone can please take a look at the reset password function that would be really helpful because this function just makes me want to cry at this point.