PLEASE READ DESCRIPTION!
Hey everyone. So there is nothing new in this edition I am posting. However, this version has NO BOTTOM NAVIGATION BAR that you saw on the home page. It is now clean. Personally, I think this version should be used from here on out but we definitely should discuss which version to use. I am leaving this version and the normal version we have up on Github so we can agree on which version to use.